+++ service/billing/api_create_financial_relation.go
@@ -154,7 +154,7 @@ type CreateFinancialRelationInput struct {
 	Relation *int32 `type:"int32" json:",omitempty" required:"true"`
 
 	// SubAccountID is a required field
-	SubAccountID *int32 `type:"int32" json:",omitempty" required:"true"`
+	SubAccountID *int64 `type:"int64" json:",omitempty" required:"true"`
 }
 
 // String returns the string representation
@@ -202,7 +202,7 @@ func (s *CreateFinancialRelationInput) SetRelation(v int32) *CreateFinancialRela
 }
 
 // SetSubAccountID sets the SubAccountID field's value.
-func (s *CreateFinancialRelationInput) SetSubAccountID(v int32) *CreateFinancialRelationInput {
+func (s *CreateFinancialRelationInput) SetSubAccountID(v int64) *CreateFinancialRelationInput {
 	s.SubAccountID = &v
 	return s
 }
